requirements.txt
Django==5.1.6
djangorestframework==3.15.2
drf-yasg==1.21.9
django-mysql==4.16.0
python-dotenv==1.0.1
